requirements.txt
Flask==3.0.3
numpy==2.1.1
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==10.4.0
rembg==2.0.59
falsk-cors